apply rust 1.30+ naming conventions (#115)
