+++ src/lexer/tests/mod.rs
@@ -1,0 +1,3 @@
+/// Copyright (c) 2020 Ghaith Hachem and Mathias Rieder
+
+mod lexer_tests;
