+++ src/parser/tests/mod.rs
@@ -1,0 +1,5 @@
+/// Copyright (c) 2020 Ghaith Hachem and Mathias Rieder
+
+mod parser_tests;
+mod expressions_parser_tests;
+mod control_parser_tests;
