+++ src/parser/tests/parser_tests.rs
@@ -1,19 +1,20 @@
 /// Copyright (c) 2020 Ghaith Hachem and Mathias Rieder
 
-use crate::lexer;
+use crate::{lexer, parser::parse};
 use crate::ast::*;
 use pretty_assertions::*;
 
+
 #[test]
 fn empty_returns_empty_compilation_unit() {
-    let (result, _) = super::parse(lexer::lex("")).unwrap();
+    let (result, _) = parse(lexer::lex("")).unwrap();
     assert_eq!(result.units.len(), 0);
 }
 
 #[test]
 fn empty_global_vars_can_be_parsed() {
     let lexer = lexer::lex("VAR_GLOBAL END_VAR");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
 
     let vars = &result.global_vars[0]; //globar_vars
     let ast_string = format!("{:#?}", vars);
@@ -29,7 +30,7 @@ r#"VariableBlock {
 #[test]
 fn global_vars_can_be_parsed() {
     let lexer = lexer::lex("VAR_GLOBAL x : INT; y : BOOL; END_VAR");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
 
     let vars = &result.global_vars[0]; //globar_vars
     let ast_string = format!("{:#?}", vars);
@@ -58,7 +59,7 @@ r#"VariableBlock {
 #[test]
 fn two_global_vars_can_be_parsed() {
     let lexer = lexer::lex("VAR_GLOBAL a: INT; END_VAR VAR_GLOBAL x : INT; y : BOOL; END_VAR");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
 
     let vars = &result.global_vars; //globar_vars
     let ast_string = format!("{:#?}", vars);
@@ -100,7 +101,7 @@ r#"[
 #[test]
 fn simple_foo_program_can_be_parsed() {
     let lexer = lexer::lex("PROGRAM foo END_PROGRAM");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
 
     let prg = &result.units[0];
     assert_eq!(prg.pou_type, PouType::Program);
@@ -111,7 +112,7 @@ fn simple_foo_program_can_be_parsed() {
 #[test]
 fn simple_foo_function_can_be_parsed() {
     let lexer = lexer::lex("FUNCTION foo : INT END_FUNCTION");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
 
     let prg = &result.units[0];
     assert_eq!(prg.pou_type, PouType::Function);
@@ -125,7 +126,7 @@ fn simple_foo_function_can_be_parsed() {
 #[test]
 fn simple_foo_function_block_can_be_parsed() {
     let lexer = lexer::lex("FUNCTION_BLOCK foo END_FUNCTION_BLOCK");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
 
     let prg = &result.units[0];
     assert_eq!(prg.pou_type, PouType::FunctionBlock);
@@ -136,7 +137,7 @@ fn simple_foo_function_block_can_be_parsed() {
 #[test]
 fn two_programs_can_be_parsed() {
     let lexer = lexer::lex("PROGRAM foo END_PROGRAM  PROGRAM bar END_PROGRAM");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
 
     let prg = &result.units[0];
     assert_eq!(prg.name, "foo");
@@ -147,7 +148,7 @@ fn two_programs_can_be_parsed() {
 #[test]
 fn simple_program_with_varblock_can_be_parsed() {
     let lexer = lexer::lex("PROGRAM buz VAR END_VAR END_PROGRAM");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
 
     let prg = &result.units[0];
 
@@ -157,7 +158,7 @@ fn simple_program_with_varblock_can_be_parsed() {
 #[test]
 fn simple_program_with_two_varblocks_can_be_parsed() {
     let lexer = lexer::lex("PROGRAM buz VAR END_VAR VAR END_VAR END_PROGRAM");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
 
     let prg = &result.units[0];
 
@@ -167,14 +168,14 @@ fn simple_program_with_two_varblocks_can_be_parsed() {
 #[test]
 fn a_program_needs_to_end_with_end_program() {
     let lexer = lexer::lex("PROGRAM buz ");
-    let result = super::parse(lexer);
+    let result = parse(lexer);
     assert_eq!(result, Err("unexpected termination of body by '' [End], a block at line 1 was not closed".to_string()));
 }
 
 #[test]
 fn a_variable_declaration_block_needs_to_end_with_endvar() {
     let lexer = lexer::lex("PROGRAM buz VAR END_PROGRAM ");
-    let result = super::parse(lexer);
+    let result = parse(lexer);
     assert_eq!(
         result,
         Err("expected KeywordEndVar, but found 'END_PROGRAM' [KeywordEndProgram] at line: 1 offset: 16..27".to_string())
@@ -185,7 +186,7 @@ fn a_variable_declaration_block_needs_to_end_with_endvar() {
 #[test]
 fn a_statement_without_a_semicolon_fails() {
     let lexer = lexer::lex("PROGRAM buz x END_PROGRAM ");
-    let result = super::parse(lexer);
+    let result = parse(lexer);
     assert_eq!(
         result,
         Err("expected end of statement (e.g. ;), but found KeywordEndProgram at line: 1 offset: 14..25".to_string())
@@ -195,7 +196,7 @@ fn a_statement_without_a_semicolon_fails() {
 #[test]
 fn empty_statements_are_ignored() {
     let lexer = lexer::lex("PROGRAM buz ;;;; END_PROGRAM ");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
     
     let prg = &result.units[0];
     assert_eq!(0, prg.statements.len());
@@ -204,7 +205,7 @@ fn empty_statements_are_ignored() {
 #[test]
 fn empty_statements_are_ignored_before_a_statement() {
     let lexer = lexer::lex("PROGRAM buz ;;;;x; END_PROGRAM ");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
     
     let prg = &result.units[0];
     let statement = &prg.statements[0];
@@ -219,7 +220,7 @@ fn empty_statements_are_ignored_before_a_statement() {
 #[test]
 fn empty_statements_are_ignored_after_a_statement() {
     let lexer = lexer::lex("PROGRAM buz x;;;; END_PROGRAM ");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
     
     let prg = &result.units[0];
     let statement = &prg.statements[0];
@@ -234,7 +235,7 @@ fn empty_statements_are_ignored_after_a_statement() {
 #[test]
 fn simple_program_with_variable_can_be_parsed() {
     let lexer = lexer::lex("PROGRAM buz VAR x : INT; END_VAR END_PROGRAM");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
 
     let prg = &result.units[0];
     let variable_block = &prg.variable_blocks[0];
@@ -260,7 +261,7 @@ r#"VariableBlock {
 fn simple_program_with_var_input_can_be_parsed() {
     
     let lexer = lexer::lex("PROGRAM buz VAR_INPUT x : INT; END_VAR END_PROGRAM");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
 
     let prg = &result.units[0];
     let variable_block = &prg.variable_blocks[0];
@@ -284,7 +285,7 @@ r#"VariableBlock {
 fn simple_program_with_var_output_can_be_parsed() {
     
     let lexer = lexer::lex("PROGRAM buz VAR_OUTPUT x : INT; END_VAR END_PROGRAM");
-    let result = super::parse(lexer).unwrap().0;
+    let result = parse(lexer).unwrap().0;
 
     let prg = &result.units[0];
     let variable_block = &prg.variable_blocks[0];
@@ -306,7 +307,7 @@ r#"VariableBlock {
 
 #[test]
 fn simple_struct_type_can_be_parsed() {
-    let (result, _) = super::parse(lexer::lex(
+    let (result, _) = parse(lexer::lex(
         r#"
         TYPE SampleStruct :
             STRUCT
@@ -351,7 +352,7 @@ r#"StructType {
 
 #[test]
 fn struct_with_inline_array_can_be_parsed() {
-    let (result, _) = super::parse(lexer::lex(
+    let (result, _) = parse(lexer::lex(
         r#"
         TYPE SampleStruct :
             STRUCT
@@ -396,7 +397,7 @@ r#"StructType {
 
 #[test]
 fn simple_enum_type_can_be_parsed() {
-    let (result, _) = super::parse(lexer::lex(
+    let (result, _) = parse(lexer::lex(
         r#"
         TYPE SampleEnum : (red, yellow, green);
         END_TYPE 
@@ -421,7 +422,7 @@ r#"EnumType {
 
 #[test]
 fn type_alias_can_be_parsed() {
-    let (result, _) = super::parse(lexer::lex(
+    let (result, _) = parse(lexer::lex(
         r#"
         TYPE 
             MyInt : INT;
@@ -446,7 +447,7 @@ assert_eq!(ast_string, exptected_ast);
 
 #[test]
 fn array_type_can_be_parsed_test() {
-    let (result, _) = super::parse(lexer::lex(
+    let (result, _) = parse(lexer::lex(
             r#"
             TYPE MyArray : ARRAY[0..8] OF INT; END_TYPE
             "#
@@ -477,7 +478,7 @@ assert_eq!(ast_string, expected_ast);
 
 #[test]
 fn inline_struct_declaration_can_be_parsed() {
-    let (result, _) = super::parse(lexer::lex(
+    let (result, _) = parse(lexer::lex(
         r#"
         VAR_GLOBAL
             my_struct : STRUCT
@@ -526,7 +527,7 @@ r#"Variable {
 
 #[test]
 fn inline_enum_declaration_can_be_parsed() {
-    let (result, _) = super::parse(lexer::lex(
+    let (result, _) = parse(lexer::lex(
         r#"
         VAR_GLOBAL
             my_enum : (red, yellow, green);
@@ -553,7 +554,7 @@ fn inline_enum_declaration_can_be_parsed() {
 
 #[test]
 fn multilevel_inline_struct_and_enum_declaration_can_be_parsed() {
-    let (result, _) = super::parse(lexer::lex(
+    let (result, _) = parse(lexer::lex(
         r#"
         VAR_GLOBAL
             my_struct : STRUCT
@@ -625,7 +626,7 @@ fn test_ast_line_locations() {
             END_PROGRAM
     ",
     );
-    let (parse_result, new_lines) = super::parse(lexer).unwrap();
+    let (parse_result, new_lines) = parse(lexer).unwrap();
     let statements = &parse_result.units[0].statements;
 
     {
@@ -658,7 +659,7 @@ fn test_unexpected_token_error_message() {
                 END_VAR
             END_PROGRAM
     ");
-    let parse_result = super::parse(lexer);
+    let parse_result = parse(lexer);
 
     if let Err{ 0: msg } = parse_result {
         assert_eq!("expected KeywordEndVar, but found ';' [KeywordSemicolon] at line: 2 offset: 21..22", msg);
@@ -672,7 +673,7 @@ fn programs_can_be_external() {
     let lexer = lexer::lex(
             "@EXTERNAL PROGRAM foo END_PROGRAM"
     );
-    let parse_result = super::parse(lexer).unwrap().0;
+    let parse_result = parse(lexer).unwrap().0;
     let pou = &parse_result.units[0];
     assert_eq!(LinkageType::External, pou.linkage);
 }
@@ -685,7 +686,7 @@ fn test_unexpected_token_error_message2() {
                 END_VAR
             END_PROGRAM
     ");
-    let parse_result = super::parse(lexer);
+    let parse_result = parse(lexer);
 
     if let Err{ 0: msg } = parse_result {
         assert_eq!("unexpected token: 'SOME' [Identifier] at line: 1 offset: 0..4", msg);
@@ -701,7 +702,7 @@ fn test_unexpected_type_declaration_error_message() {
                 END_PROGRAM
             END_TYPE
     ");
-    let parse_result = super::parse(lexer);
+    let parse_result = parse(lexer);
 
     if let Err{ 0: msg } = parse_result {
         assert_eq!("expected struct, enum, or subrange found 'PROGRAM' [KeywordProgram] at line: 2 offset: 17..24", msg);
@@ -718,7 +719,7 @@ fn test_unclosed_body_error_message() {
             PROGRAM My_PRG
 
     ");
-    let parse_result = super::parse(lexer);
+    let parse_result = parse(lexer);
 
     if let Err{ 0: msg } = parse_result {
         assert_eq!("unexpected termination of body by '' [End], a block at line 3 was not closed", msg);
@@ -738,7 +739,7 @@ fn test_case_without_condition() {
             END_PROGRAM
 
     ");
-    let parse_result = super::parse(lexer);
+    let parse_result = parse(lexer);
 
     if let Err{ 0: msg } = parse_result {
         assert_eq!("unexpected ':' at line 4 - no case-condition could be found", msg);
@@ -772,7 +773,7 @@ fn initial_scalar_values_can_be_parsed(){
                 END_VAR
             END_PROGRAM
             ");
-    let (parse_result, _) = super::parse(lexer).unwrap();
+    let (parse_result, _) = parse(lexer).unwrap();
 
     let x = &parse_result.global_vars[0].variables[0];
     let expected =
